refactors neumorph logo component as a molecule
