+++ src/main/components/UI/organisms/DropFile/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { confirmSize } from "../../../../utilities/helpers/files";
 import { Button } from "../../atoms/Button";
-import { NeuMorphLogo } from "../../atoms/NeumorphLogo";
+import { NeuMorphLogo } from "../../molecules/NeumorphLogo";
 import styles from "./index.module.css";
 
 interface DropFileProps {
